requirements.txt

ultralytics
PyYAML
mlflow
dvc[gdrive]
dvc-gdrive

